Call the post and term crumbs from the associated builder classes.
This removes the post and term calls from the query classes and puts them in the builders. The old method seemed to be a leftover from an older version of the plugin. It doesn't really make sense to call a post/term builder and then the crumb when the builder itself could be doing the job.