Adds `spacing.blockGap` and `layout` support.
This commit implements the Core `layout` support, which is a prerequisite for `spacing.blockGap` to work. It adds support for `innerBlock` (which this block doesn't technically have) to achieve this.

The block maintains its own content justification right now. This is primarily to avoid supporting the `space-between` justification that Core has, which is not needed for this block.

Fixes: https://github.com/x3p0-dev/x3p0-breadcrumbs/issues/6