Enable rewrite tag mapping for all post types with tags by default.
This gives us consistent handling of rewrite tag mapping where we were manually doing it for only the Core `post` post type before.

This change also removes the old-school method of auto-setting the `category` taxonomy for the `post` post type when only `%postname%` was in the permalink structure. This was a relic from when users didn't have direct access to make the decision themselves on tying a taxonomy to a post type.